requirements.txt
Flask
transformers
torch
Pillow